C-API types revisited (#1770)
JIRA: CVS-104313

+++ src/main_capi.cpp
@@ -29,10 +29,10 @@
 #include "ovms.h"  // NOLINT
 
 const char* MODEL_NAME = "dummy";
-const uint64_t MODEL_VERSION = 1;
+const int64_t MODEL_VERSION = 1;
 const char* INPUT_NAME = "b";
 constexpr size_t DIM_COUNT = 2;
-constexpr size_t SHAPE[DIM_COUNT] = {1, 10};
+constexpr int64_t SHAPE[DIM_COUNT] = {1, 10};
 
 namespace {
 volatile sig_atomic_t shutdown_request = 0;
@@ -142,8 +142,8 @@ int main(int argc, char** argv) {
     size_t bytesize = 0;
     uint32_t outputId = outputCount - 1;
     OVMS_DataType datatype = (OVMS_DataType)42;
-    const uint64_t* shape{nullptr};
-    uint32_t dimCount = 0;
+    const int64_t* shape{nullptr};
+    size_t dimCount = 0;
     OVMS_BufferType bufferType = (OVMS_BufferType)42;
     uint32_t deviceId = 42;
     const char* outputName{nullptr};
